pyproject.toml
[project]
name = "eflycode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "openai>=1.0.0",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "rich>=14.1.0",
    "toml>=0.10.2",
    "tree-sitter>=0.25.1",
    "tree-sitter-language-pack>=0.9.0",
    "tree-sitter-languages>=1.10.2",
    "typing-extensions>=4.15.0",
    "tiktoken>=0.12.0",
    "watchdog>=6.0.0",
    "wcwidth>=0.2.13",
]


[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
